change relics to generic components
