+++ src/parser.js
@@ -1,6 +1,6 @@
 var request = require('request');
 var WorldState = require('./worldstate.js');
-var RelicQuery = require('./modules/relics.js');
+var LocationQuery = require('./modules/components.js');
 
 var MAX_CACHED_TIME = process.env.WORLDSTATE_CACHE_LENGTH || 300000;
 
@@ -558,9 +558,9 @@ Parser.prototype.getDarkSectorsString = function(callback) {
   });
 }
 
-//Relics
-Parser.getRelicFromQuery = function(query, callback){
-  new RelicQuery(query, function(err, parts){
+//Component Query lookup
+Parser.getComponentFromQuery = function(query, callback){
+  new LocationQuery(query, function(err, parts){
     if(err){
       return callback(err);
     }
